requirements.txt
Django>=2.2.0
social-auth-app-django>=4.0.0
python-jose>=3.2.0